requirements.txt
absl-py==2.4.0
alembic==1.18.4
asttokens==3.0.1
astunparse==1.6.3
bleach==6.3.0
certifi==2026.2.25
charset-normalizer==3.4.5
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.3.0
gast==0.7.0
google-pasta==0.2.0
greenlet==3.3.2
grpcio==1.80.0
h5py==3.14.0
idna==3.11
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kaggle==2.0.0
kagglesdk==0.1.16
keras==3.13.2
kiwisolver==1.5.0
libclang==18.1.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
missingno==0.5.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.3
opt_einsum==3.4.0
optree==0.19.0
optuna==4.8.0
packaging==26.0
pandas==3.0.1
parso==0.8.6
pillow==12.1.1
platformdirs==4.9.4
prompt_toolkit==3.0.52
protobuf==7.34.0
psutil==7.2.2
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
rich==14.3.3
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==81.0.0
six==1.17.0
SQLAlchemy==2.0.48
stack-data==0.6.3
sympy==1.14.0
tensorflow==2.21.0
termcolor==3.3.0
text-unidecode==1.3
threadpoolctl==3.6.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
webencodings==0.5.1
wheel==0.46.3
wrapt==2.1.2
numpy
pandas
matplotlib
scikit-learn
torch
pyyaml
tensorflow>=2.10.0
